requirements.txt
fastapi
uvicorn[standard]
pyodbc
openai
python-dotenv
